// src/protos/api_pb.d.ts
export class Organization extends jspb.Message {
  getExternalUuid(): string;
  setExternalUuid(value: string): Organization;
  getName(): string;
  setName(value: string): Organization;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): Organization.AsObject;
  static toObject(
    includeInstance: boolean,
    msg: Organization
  ): Organization.AsObject;
  static extensions: { [key: number]: jspb.ExtensionFieldInfo<jspb.Message> };
  static extensionsBinary: {
    [key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>;
  };
  static serializeBinaryToWriter(
    message: Organization,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): Organization;
  static deserializeBinaryFromReader(
    message: Organization,
    reader: jspb.BinaryReader
  ): Organization;
}

export namespace Organization {
  export type AsObject = {
    externalUuid: string;
    name: string;
  };
}

export class School extends jspb.Message {
  getExternalUuid(): string;
  setExternalUuid(value: string): School;
  getExternalOrganizationUuid(): string;
  setExternalOrganizationUuid(value: string): School;
  getName(): string;
  setName(value: string): School;
  getShortCode(): string;
  setShortCode(value: string): School;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): School.AsObject;
  static toObject(includeInstance: boolean, msg: School): School.AsObject;
  static extensions: { [key: number]: jspb.ExtensionFieldInfo<jspb.Message> };
  static extensionsBinary: {
    [key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>;
  };
  static serializeBinaryToWriter(
    message: School,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): School;
  static deserializeBinaryFromReader(
    message: School,
    reader: jspb.BinaryReader
  ): School;
}

export namespace School {
  export type AsObject = {
    externalUuid: string;
    externalOrganizationUuid: string;
    name: string;
    shortCode: string;
  };
}

export class Class extends jspb.Message {
  getExternalUuid(): string;
  setExternalUuid(value: string): Class;
  getExternalOrganizationUuid(): string;
  setExternalOrganizationUuid(value: string): Class;
  getName(): string;
  setName(value: string): Class;
  getExternalSchoolUuid(): string;
  setExternalSchoolUuid(value: string): Class;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): Class.AsObject;
  static toObject(includeInstance: boolean, msg: Class): Class.AsObject;
  static extensions: { [key: number]: jspb.ExtensionFieldInfo<jspb.Message> };
  static extensionsBinary: {
    [key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>;
  };
  static serializeBinaryToWriter(
    message: Class,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): Class;
  static deserializeBinaryFromReader(
    message: Class,
    reader: jspb.BinaryReader
  ): Class;
}

export namespace Class {
  export type AsObject = {
    externalUuid: string;
    externalOrganizationUuid: string;
    name: string;
    externalSchoolUuid: string;
  };
}

export class User extends jspb.Message {
  getExternalUuid(): string;
  setExternalUuid(value: string): User;
  getExternalOrganizationUuid(): string;
  setExternalOrganizationUuid(value: string): User;
  getEmail(): string;
  setEmail(value: string): User;
  getPhone(): string;
  setPhone(value: string): User;
  getUsername(): string;
  setUsername(value: string): User;
  getGivenName(): string;
  setGivenName(value: string): User;
  getFamilyName(): string;
  setFamilyName(value: string): User;
  getGender(): Gender;
  setGender(value: Gender): User;
  getDateOfBirth(): string;
  setDateOfBirth(value: string): User;
  clearRoleIdentifiersList(): void;
  getRoleIdentifiersList(): Array<string>;
  setRoleIdentifiersList(value: Array<string>): User;
  addRoleIdentifiers(value: string, index?: number): string;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): User.AsObject;
  static toObject(includeInstance: boolean, msg: User): User.AsObject;
  static extensions: { [key: number]: jspb.ExtensionFieldInfo<jspb.Message> };
  static extensionsBinary: {
    [key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>;
  };
  static serializeBinaryToWriter(
    message: User,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): User;
  static deserializeBinaryFromReader(
    message: User,
    reader: jspb.BinaryReader
  ): User;
}

export namespace User {
  export type AsObject = {
    externalUuid: string;
    externalOrganizationUuid: string;
    email: string;
    phone: string;
    username: string;
    givenName: string;
    familyName: string;
    gender: Gender;
    dateOfBirth: string;
    roleIdentifiersList: Array<string>;
  };
}
// ...
export class Link extends jspb.Message {
  hasAddUsersToOrganization(): boolean;
  clearAddUsersToOrganization(): void;
  getAddUsersToOrganization(): AddUsersToOrganization | undefined;
  setAddUsersToOrganization(value?: AddUsersToOrganization): Link;

  hasAddOrganizationRolesToUser(): boolean;
  clearAddOrganizationRolesToUser(): void;
  getAddOrganizationRolesToUser(): AddOrganizationRolesToUser | undefined;
  setAddOrganizationRolesToUser(value?: AddOrganizationRolesToUser): Link;

  hasAddUsersToSchool(): boolean;
  clearAddUsersToSchool(): void;
  getAddUsersToSchool(): AddUsersToSchool | undefined;
  setAddUsersToSchool(value?: AddUsersToSchool): Link;

  hasAddUsersToClass(): boolean;
  clearAddUsersToClass(): void;
  getAddUsersToClass(): AddUsersToClass | undefined;
  setAddUsersToClass(value?: AddUsersToClass): Link;

  hasAddProgramsToSchool(): boolean;
  clearAddProgramsToSchool(): void;
  getAddProgramsToSchool(): AddProgramsToSchool | undefined;
  setAddProgramsToSchool(value?: AddProgramsToSchool): Link;

  hasAddProgramsToClass(): boolean;
  clearAddProgramsToClass(): void;
  getAddProgramsToClass(): AddProgramsToClass | undefined;
  setAddProgramsToClass(value?: AddProgramsToClass): Link;

  hasAddClassesToSchool(): boolean;
  clearAddClassesToSchool(): void;
  getAddClassesToSchool(): AddClassesToSchool | undefined;
  setAddClassesToSchool(value?: AddClassesToSchool): Link;

  getLinkCase(): Link.LinkCase;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): Link.AsObject;
  static toObject(includeInstance: boolean, msg: Link): Link.AsObject;
  static extensions: { [key: number]: jspb.ExtensionFieldInfo<jspb.Message> };
  static extensionsBinary: {
    [key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>;
  };
  static serializeBinaryToWriter(
    message: Link,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): Link;
  static deserializeBinaryFromReader(
    message: Link,
    reader: jspb.BinaryReader
  ): Link;
}

export namespace Link {
  export type AsObject = {
    addUsersToOrganization?: AddUsersToOrganization.AsObject;
    addOrganizationRolesToUser?: AddOrganizationRolesToUser.AsObject;
    addUsersToSchool?: AddUsersToSchool.AsObject;
    addUsersToClass?: AddUsersToClass.AsObject;
    addProgramsToSchool?: AddProgramsToSchool.AsObject;
    addProgramsToClass?: AddProgramsToClass.AsObject;
    addClassesToSchool?: AddClassesToSchool.AsObject;
  };

  export enum LinkCase {
    LINK_NOT_SET = 0,
    ADD_USERS_TO_ORGANIZATION = 1,
    ADD_ORGANIZATION_ROLES_TO_USER = 2,
    ADD_USERS_TO_SCHOOL = 3,
    ADD_USERS_TO_CLASS = 4,
    ADD_PROGRAMS_TO_SCHOOL = 5,
    ADD_PROGRAMS_TO_CLASS = 6,
    ADD_CLASSES_TO_SCHOOL = 7,
  }
}

export class OnboardingRequest extends jspb.Message {
  hasRequestId(): boolean;
  clearRequestId(): void;
  getRequestId(): RequestMetadata | undefined;
  setRequestId(value?: RequestMetadata): OnboardingRequest;
  getAction(): Action;
  setAction(value: Action): OnboardingRequest;

  hasLinkEntities(): boolean;
  clearLinkEntities(): void;
  getLinkEntities(): Link | undefined;
  setLinkEntities(value?: Link): OnboardingRequest;

  hasOrganization(): boolean;
  clearOrganization(): void;
  getOrganization(): Organization | undefined;
  setOrganization(value?: Organization): OnboardingRequest;

  hasSchool(): boolean;
  clearSchool(): void;
  getSchool(): School | undefined;
  setSchool(value?: School): OnboardingRequest;

  hasClass(): boolean;
  clearClass(): void;
  getClass(): Class | undefined;
  setClass(value?: Class): OnboardingRequest;

  hasUser(): boolean;
  clearUser(): void;
  getUser(): User | undefined;
  setUser(value?: User): OnboardingRequest;

  getPayloadCase(): OnboardingRequest.PayloadCase;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): OnboardingRequest.AsObject;
  static toObject(
    includeInstance: boolean,
    msg: OnboardingRequest
  ): OnboardingRequest.AsObject;
  static extensions: { [key: number]: jspb.ExtensionFieldInfo<jspb.Message> };
  static extensionsBinary: {
    [key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>;
  };
  static serializeBinaryToWriter(
    message: OnboardingRequest,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): OnboardingRequest;
  static deserializeBinaryFromReader(
    message: OnboardingRequest,
    reader: jspb.BinaryReader
  ): OnboardingRequest;
}

export namespace OnboardingRequest {
  export type AsObject = {
    requestId?: RequestMetadata.AsObject;
    action: Action;
    linkEntities?: Link.AsObject;
    organization?: Organization.AsObject;
    school?: School.AsObject;
    pb_class?: Class.AsObject;
    user?: User.AsObject;
  };

  export enum PayloadCase {
    PAYLOAD_NOT_SET = 0,
    LINK_ENTITIES = 16,
    ORGANIZATION = 17,
    SCHOOL = 18,
    CLASS = 19,
    USER = 20,
  }
}

export class BatchOnboarding extends jspb.Message {
  clearRequestsList(): void;
  getRequestsList(): Array<OnboardingRequest>;
  setRequestsList(value: Array<OnboardingRequest>): BatchOnboarding;
  addRequests(value?: OnboardingRequest, index?: number): OnboardingRequest;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): BatchOnboarding.AsObject;
  static toObject(
    includeInstance: boolean,
    msg: BatchOnboarding
  ): BatchOnboarding.AsObject;
  static extensions: { [key: number]: jspb.ExtensionFieldInfo<jspb.Message> };
  static extensionsBinary: {
    [key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>;
  };
  static serializeBinaryToWriter(
    message: BatchOnboarding,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): BatchOnboarding;
  static deserializeBinaryFromReader(
    message: BatchOnboarding,
    reader: jspb.BinaryReader
  ): BatchOnboarding;
}

export namespace BatchOnboarding {
  export type AsObject = {
    requestsList: Array<OnboardingRequest.AsObject>;
  };
}
// ...
export class RequestMetadata extends jspb.Message {
  getId(): string;
  setId(value: string): RequestMetadata;
  getN(): string;
  setN(value: string): RequestMetadata;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): RequestMetadata.AsObject;
  static toObject(
    includeInstance: boolean,
    msg: RequestMetadata
  ): RequestMetadata.AsObject;
  static extensions: { [key: number]: jspb.ExtensionFieldInfo<jspb.Message> };
  static extensionsBinary: {
    [key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>;
  };
  static serializeBinaryToWriter(
    message: RequestMetadata,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): RequestMetadata;
  static deserializeBinaryFromReader(
    message: RequestMetadata,
    reader: jspb.BinaryReader
  ): RequestMetadata;
}

export namespace RequestMetadata {
  export type AsObject = {
    id: string;
    n: string;
  };
}
// ...
export enum Action {
  CREATE = 0,
}